drivers: power/supply: Fix format specifier warning

+++ drivers/power/supply/qcom/smb5-lib.c
@@ -6858,7 +6858,7 @@ static void smblib_handle_hvdcp_3p0_auth_done(struct smb_charger *chg,
 				!timer_pending(&chg->apsd_timer)) {
 			smblib_dbg(chg, PR_MISC,
 				"APSD Extented timer started at %lld\n",
-				jiffies_to_msecs(jiffies));
+				(unsigned long long)jiffies_to_msecs(jiffies));
 
 			mod_timer(&chg->apsd_timer,
 				msecs_to_jiffies(APSD_EXTENDED_TIMEOUT_MS)
@@ -8859,7 +8859,7 @@ static void apsd_timer_cb(unsigned long data)
 	struct smb_charger *chg = (struct smb_charger *)data;
 
 	smblib_dbg(chg, PR_MISC, "APSD Extented timer timeout at %lld\n",
-			jiffies_to_msecs(jiffies));
+			(unsigned long long)jiffies_to_msecs(jiffies));
 
 	chg->apsd_ext_timeout = true;
 }
